+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Unreleased
 * Generate CHANGELOG entry in the terra-dev-site docs
 * Added ChangeLog.jsx template for terra-dev-site docs
 
+### Changed
+* Added .module extension to scss file
+
 1.11.0 - (June 26, 2018)
 ------------------
 ### Changed